backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.2
django-cors-headers>=4.3
drf-spectacular>=0.27
django-extensions>=3.2
psycopg2-binary>=2.9
gunicorn>=21.2
uvicorn[standard]>=0.23
celery>=5.3
redis>=5.0
boto3>=1.28
python-dotenv>=1.0
pillow>=10.0
WeasyPrint>=60.0
tablib[xlsx]>=3.5
reportlab>=4.0
requests>=2.31
cryptography>=41.0
sentry-sdk[django]>=1.38.0
prometheus-client>=0.19.0
django-prometheus>=2.3.1

